+++ app/src/main/java/de/danoeh/antennapod/ui/screen/drawer/DrawerPreferencesDialog.java
@@ -4,8 +4,8 @@ import android.content.Context;
 import androidx.annotation.NonNull;
 import de.danoeh.antennapod.R;
 import de.danoeh.antennapod.storage.preferences.UserPreferences;
-import de.danoeh.antennapod.ui.screen.preferences.ReorderDialog;
-import de.danoeh.antennapod.ui.screen.preferences.ReorderDialogItem;
+import de.danoeh.antennapod.ui.preferences.screen.ReorderDialog;
+import de.danoeh.antennapod.ui.preferences.screen.ReorderDialogItem;
 
 import java.util.ArrayList;
 import java.util.Collections;
