+++ app/src/main/java/de/danoeh/antennapod/ui/screen/feed/RenameFeedDialog.java
@@ -11,7 +11,7 @@ import de.danoeh.antennapod.R;
 import de.danoeh.antennapod.storage.database.NavDrawerData;
 import de.danoeh.antennapod.model.feed.Feed;
 import de.danoeh.antennapod.storage.database.DBWriter;
-import de.danoeh.antennapod.databinding.EditTextDialogBinding;
+import de.danoeh.antennapod.ui.common.databinding.EditTextDialogBinding;
 import de.danoeh.antennapod.model.feed.FeedPreferences;
 
 public class RenameFeedDialog {
