+++ app/src/main/java/de/danoeh/antennapod/ui/screen/subscriptions/FeedMultiSelectActionHandler.java
@@ -21,8 +21,8 @@ import de.danoeh.antennapod.ui.screen.feed.RemoveFeedDialog;
 import de.danoeh.antennapod.ui.screen.feed.preferences.TagSettingsDialog;
 import de.danoeh.antennapod.model.feed.Feed;
 import de.danoeh.antennapod.model.feed.FeedPreferences;
-import de.danoeh.antennapod.ui.screen.preferences.PreferenceListDialog;
-import de.danoeh.antennapod.ui.screen.preferences.PreferenceSwitchDialog;
+import de.danoeh.antennapod.ui.preferences.screen.PreferenceListDialog;
+import de.danoeh.antennapod.ui.preferences.screen.PreferenceSwitchDialog;
 import io.reactivex.rxjava3.android.schedulers.AndroidSchedulers;
 import io.reactivex.rxjava3.core.Observable;
 import io.reactivex.rxjava3.schedulers.Schedulers;
