+++ ui/common/src/main/res/layout/edit_text_dialog.xml
