+++ ui/preferences/src/main/java/de/danoeh/antennapod/ui/preferences/screen/AutomaticDeletionPreferencesFragment.java
@@ -1,13 +1,13 @@
-package de.danoeh.antennapod.ui.screen.preferences;
+package de.danoeh.antennapod.ui.preferences.screen;
 
 import android.content.res.Resources;
 import android.os.Bundle;
+import androidx.appcompat.app.AppCompatActivity;
 import androidx.preference.ListPreference;
 import androidx.preference.TwoStatePreference;
 import com.google.android.material.dialog.MaterialAlertDialogBuilder;
-import de.danoeh.antennapod.R;
+import de.danoeh.antennapod.ui.preferences.R;
 import de.danoeh.antennapod.storage.preferences.UserPreferences;
-import de.danoeh.antennapod.ui.preferences.screen.AnimatedPreferenceFragment;
 
 
 public class AutomaticDeletionPreferencesFragment extends AnimatedPreferenceFragment {
@@ -25,7 +25,7 @@ public class AutomaticDeletionPreferencesFragment extends AnimatedPreferenceFrag
     @Override
     public void onStart() {
         super.onStart();
-        ((PreferenceActivity) getActivity()).getSupportActionBar().setTitle(R.string.pref_auto_delete_title);
+        ((AppCompatActivity) getActivity()).getSupportActionBar().setTitle(R.string.pref_auto_delete_title);
     }
 
     private void checkItemVisibility(boolean autoDeleteEnabled) {
