+++ ui/preferences/src/main/java/de/danoeh/antennapod/ui/preferences/screen/ReorderDialogItem.java
@@ -1,4 +1,4 @@
-package de.danoeh.antennapod.ui.screen.preferences;
+package de.danoeh.antennapod.ui.preferences.screen;
 
 public final class ReorderDialogItem {
     public enum ViewType {
