+++ ui/preferences/src/main/java/de/danoeh/antennapod/ui/preferences/screen/ReorderItemTouchCallback.java
@@ -1,4 +1,4 @@
-package de.danoeh.antennapod.ui.screen.preferences;
+package de.danoeh.antennapod.ui.preferences.screen;
 
 import androidx.annotation.NonNull;
 import androidx.recyclerview.widget.ItemTouchHelper;
