+++ ui/preferences/src/main/res/layout/reorder_dialog.xml
