+++ ui/preferences/src/main/res/layout/reorder_dialog_entry.xml
